feat: Add lnurlAuth method to KeyringHd and Vault classes
